+++ pyprot/pdbio.py
@@ -23,7 +23,6 @@ class PdbIO(object):
           `True` if file was successfully written.
 
         """
-
         try:
             with open(dest, 'w') as out:
                 for line in self.cont:
@@ -34,3 +33,34 @@ class PdbIO(object):
             success = False
 
         return success
+
+        
+    def fetch_rcsb(self, pdb_code):
+        """
+        Fetches PDB file contents from rcsb.org.
+
+        Parameters
+        ----------
+        
+        pdb_code : `str`.
+          A 4-letter PDB code, e.g., `"3eiy"`
+        
+        Returns
+        ----------
+
+        pdb_cont : `list`.
+          List of PDB file contents after where list item is a `str` of
+          PDB file contents.
+            
+        """   
+        pdb_cont = []
+        try:
+            response = urllib.request.urlopen('http://www.rcsb.org/pdb/files/%s.pdb' %pdb_code.lower())
+            dat = response.read().decode('utf-8')
+            pdb_cont = [row.strip() for row in dat.split('\n') if row.strip()]
+        except urllib.request.HTTPError as e:
+            print('HTTP Error %s' %e.code)
+        except urllib.request.URLError as e:
+            print('URL Error %s' %e.args) 
+        
+        return pdb_cont