+++ pyprot/statsbasic.py
@@ -9,13 +9,21 @@ Functions for basic statistics.
 def mode(data_list, print_out=False):
     """
     Calculates the mode of items in a list.
+    
+    Parameters
+    ----------
 
-    Keyword arguments:
-       data_list (list): list of items of any type
-       print_out (bool): if True, prints results to screen
+    data_list : `list`.
+      List of items of any type.
+      
+    print_out : `bool` (default=`False`).
+      If `True`, prints results to screen.
 
-    Returns:
-       A list of item(s) with highest mode.
+    Returns
+    ----------
+    
+    mode : `list`.  
+      A list of item(s) with highest mode.
        
     """
     freq = dict()
@@ -38,17 +46,48 @@ def mode(data_list, print_out=False):
 
 
 def mean(data_list):
-    """ Returns the mean of numbers in a list. """
+    """ 
+    Returns the sample mean of numbers in a list. 
+    
+    Parameters
+    ----------
+
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+
+    Returns
+    ----------
+    
+    sample_mean : `float`.  
+      Sample mean as `float`.
+    
+    """
     if not data_list:
         return None
     total = 0
     for ele in data_list:
         total += ele
-    return float(total)/len(data_list)
+    sample_mean = float(total)/len(data_list)
+    return sample_mean
 
 
 def median(data_list):
-    """ Returns the median of a list of numbers. """
+    """ 
+    Returns the sample median of numbers in a list. 
+    
+    Parameters
+    ----------
+
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+
+    Returns
+    ----------
+    
+    sample_median : `float`.  
+      Sample median as `float`.
+    
+    """
     if not data_list:
         return None
     sorted_data = sorted(data_list)
@@ -65,7 +104,22 @@ def median(data_list):
 
 
 def quartile1(data_list):
-    """ Returns value of the 1st quartile from a list of values. """
+    """ 
+    Returns the first quartile of numbers in a list. 
+    
+    Parameters
+    ----------
+
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+
+    Returns
+    ----------
+    
+    q1_val : `float`.  
+      First quartile as `float`.
+    
+    """
     sorted_data = sorted(data_list)
     length = len(sorted_data)
     if length % 2 != 0:
@@ -76,7 +130,22 @@ def quartile1(data_list):
 
 
 def quartile3(data_list):
-    """ Returns value of the 3rd quartile from a list of values. """
+    """ 
+    Returns the third quartile of numbers in a list. 
+    
+    Parameters
+    ----------
+
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+
+    Returns
+    ----------
+    
+    q3_val : `float`.  
+      Third quartile as `float`.
+    
+    """
     sorted_data = sorted(data_list)
     length = len(sorted_data)
     if length % 2 != 0:
@@ -87,50 +156,106 @@ def quartile3(data_list):
 
 
 def iqr(data_list):
-    """Returns the interquartile range from a list of values. """
+    """ 
+    Returns the interquartile range of numbers in a list. 
+    
+    Parameters
+    ----------
+
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+
+    Returns
+    ----------
+    
+    iqr : `float`.  
+      Interquartile range as `float`.
+    
+    """
     q1 = quartile1(data_list)
     q3 = quartile3(data_list)
     return q3 - q1
 
 
-def variance(data_list, population=True):
-    """
-    Calculates the variance from the mean of list of data.
+def variance(data_list, population=False):
+    """ 
+    Calculates the sample variance from a list of data.
+    
+    Parameters
+    ----------
 
-    Keyword arguments:
-        population: If True, calculates the sample variance
-            with Bessel's correction (n - 1) to account for higher
-            variability in sample distribution TO ESTIMATE the true population
-            standard deviation.
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+      
+    population : `bool` (default=`False`)..
+      If False, calculates the sample variance
+      with Bessel's correction (n - 1) to account for higher
+      variability in sample distribution TO ESTIMATE the true population
+      variance.
 
-    Returns:
-        The variance as a float.
-
+    Returns
+    ----------
+    
+    var : `float`.  
+      Sample variance as `float`.
+    
     """
     mean_val = sum(data_list) / float(len(data_list))
     dev = sum([((i - mean_val)**2) for i in data_list])
     if not population and len(data_list) > 2:
-        result = dev / (len(data_list) - 1)
+        var = dev / (len(data_list) - 1)
     else:
-        result = dev / len(data_list)
-    return result
+        var = dev / len(data_list)
+    return var
 
-def std_dev(data_list, population=True):
+def std_dev(data_list, population=False):
+    """ 
+    Calculates the sample standard deviation from a list of data.
+    
+    Parameters
+    ----------
+
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+      
+    population : `bool` (default=`False`).
+      If False, calculates the sample standard deviation
+      with Bessel's correction (n - 1) to account for higher
+      variability in sample distribution TO ESTIMATE the true population
+      standard deviation.
+
+    Returns
+    ----------
+    
+    stdev : `float`.  
+      Sample standard deviation as `float`.
+    
     """
-    Calculates the standard deviation from the mean in a list of data.
+    stdev = variance(data_list, population)**0.5
+    return stdev
 
-    Keyword arguments:
-        population: If True, calculates the sample standard
-            deviation with Bessel's correction (n - 1) to account for higher
-            variability in sample distribution TO ESTIMATE the true population
-            standard deviation.
-
-    Returns:
-        The standard deviation as a float.
-
-     """
-    return variance(data_list, population)**0.5
 
 def std_err(data_list, population=False):
-    """ Returns standard error s/sqrt(n). """
-    return round(std_dev(data_list, population)/len(data_list)**0.5, 4)
+    """ 
+    Calculates the sample standard error from a list of data.
+    
+    Parameters
+    ----------
+
+    data_list : `list`.
+      List of numeric items (`int` or `float`).
+      
+    population : `bool` (default=`False`).
+      If False, calculates the sample standard error
+      with Bessel's correction (n - 1) to account for higher
+      variability in sample distribution TO ESTIMATE the true population
+      standard error.
+
+    Returns
+    ----------
+    
+    sterr : `float`.  
+      Sample standard error as `float`.
+    
+    """
+    return std_dev(data_list, population)/len(data_list)**0.5