+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+- New `PdbIO` class for `Pdb` base class.
+
+## 1.0.2
+- Added distance filter functions for MOL2 files.
+
 ## 1.0.1
 - Added intermolecular MOL2 functional group filter functions and script.